refactor: add enum pakage
