+++ src/main/java/com/kurierfree/server/domain/user/domain/DisabledStudent.java
@@ -1,6 +1,9 @@
 package com.kurierfree.server.domain.user.domain;
 
 import com.kurierfree.server.domain.semester.domain.Semester;
+import com.kurierfree.server.domain.user.domain.enums.DisabilityType;
+import com.kurierfree.server.domain.user.domain.enums.Gender;
+import com.kurierfree.server.domain.user.domain.enums.Role;
 import jakarta.persistence.*;
 import lombok.AccessLevel;
 import lombok.NoArgsConstructor;
