--- src/main/java/com/kurierfree/server/domain/user/domain/Gender.java
@@ -1,5 +1,0 @@
-package com.kurierfree.server.domain.user.domain;
-
-public enum Gender {
-    MALE, FEMALE
-}
