+++ src/main/java/com/kurierfree/server/domain/user/domain/enums/Gender.java
@@ -1,0 +1,5 @@
+package com.kurierfree.server.domain.user.domain.enums;
+
+public enum Gender {
+    MALE, FEMALE
+}
