+++ src/main/java/com/kurierfree/server/domain/user/dto/response/UserResponse.java
@@ -1,6 +1,6 @@
 package com.kurierfree.server.domain.user.dto.response;
 
-import com.kurierfree.server.domain.user.domain.Role;
+import com.kurierfree.server.domain.user.domain.enums.Role;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
